rework physics context again, satisfy eslint
