+++ src/components/index.ts
@@ -2,6 +2,9 @@ export { EntropicController } from "./entropic-controller"
 export { Entropy, useEntropyContext } from "../contexts/entropy"
 export { Field } from "./field"
 export { Matter } from "./matter"
-export { Physics, usePhysics } from "../contexts/physics"
+export {
+  PhysicsProvider as Physics,
+  usePhysics,
+} from "../contexts/physics-provider"
 export { Particle } from "./particle"
 export { Title } from "./title"
