+++ src/universe.tsx
@@ -1,19 +1,20 @@
 import "./universe.css"
 
-import { Physics, Entropy, Field, EntropicController, Title, Matter } from "./components"
+import { Entropy, Field, EntropicController, Title, Matter } from "./components"
+import { PhysicsProvider } from "./contexts"
 
 export default function Universe() {
   return (
     <div className="universe cmbr">
       <Title />
-      <Physics>
+      <PhysicsProvider>
         <Entropy>
           <Field>
             <Matter />
           </Field>
           <EntropicController />
         </Entropy>
-      </Physics>
+      </PhysicsProvider>
     </div>
   )
 }
